requirements.txt
# Core Streamlit dependencies
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0

# Data science and visualization
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Chemistry and molecular handling
rdkit-pypi>=2022.9.5
mordred>=1.2.0

# File handling and utilities
openpyxl>=3.1.0
xlsxwriter>=3.1.0
toml>=0.10.2
pyyaml>=6.0

# Machine learning (lightweight versions)
torch>=2.0.0
torchvision>=0.15.0

# Optional: For advanced molecular descriptors
# Note: Some REINVENT4 dependencies may need to be installed separately
# depending on your specific deployment environment
admet-ai>=0.3.3
